Splitting up the playground Route
so that there's 1 Route per target
